+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -20,7 +20,7 @@ import { learnMore } from '../utils/link';
 
 const debug = require('debug')('expo:export:generateStaticRoutes') as typeof console.log;
 
-type Options = { outputDir: string; minify: boolean };
+type Options = { outputDir: string; minify: boolean; assetPrefix: string };
 
 /** @private */
 export async function unstable_exportStaticAsync(projectRoot: string, options: Options) {
@@ -91,9 +91,12 @@ export async function getFilesToExportFromServerAsync(
 export async function exportFromServerAsync(
   projectRoot: string,
   devServerManager: DevServerManager,
-  { outputDir, minify }: Options
+  { outputDir, assetPrefix, minify }: Options
 ): Promise<void> {
-  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, outputDir);
+  const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, {
+    assetPrefix,
+    outputDir,
+  });
 
   const devServer = devServerManager.getDefaultDevServer();
   assert(devServer instanceof MetroBundlerDevServer);
@@ -116,6 +119,7 @@ export async function exportFromServerAsync(
         mode: 'production',
         resources,
         template,
+        assetPrefix,
       });
 
       if (injectFaviconTag) {