+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -86,10 +86,13 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     resources,
     template,
     devBundleUrl,
+    assetPrefix,
   }: {
     mode: 'development' | 'production';
     resources: SerialAsset[];
     template: string;
+    /** asset prefix used for deploying to non-standard origins like GitHub pages. */
+    assetPrefix: string;
     devBundleUrl?: string;
   }): Promise<string> {
     if (!resources) {
@@ -99,6 +102,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     return htmlFromSerialAssets(resources, {
       dev: isDev,
       template,
+      assetPrefix,
       bundleUrl: isDev ? devBundleUrl : undefined,
     });
   }
@@ -221,9 +225,11 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     {
       mode,
       minify = mode !== 'development',
+      assetPrefix,
     }: {
       mode: 'development' | 'production';
       minify?: boolean;
+      assetPrefix: string;
     }
   ) {
     const devBundleUrlPathname = createBundleUrlPath({
@@ -259,6 +265,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       resources,
       template: staticHtml,
       devBundleUrl: devBundleUrlPathname,
+      assetPrefix,
     });
     return {
       content,
@@ -389,6 +396,8 @@ export class MetroBundlerDevServer extends BundlerDevServer {
           try {
             const { content } = await this.getStaticPageAsync(req.url, {
               mode: options.mode ?? 'development',
+              // Asset prefix is not supported in development.
+              assetPrefix: '',
             });
 
             res.setHeader('Content-Type', 'text/html');
@@ -525,7 +534,18 @@ export function getDeepLinkHandler(projectRoot: string): DeepLinkHandler {
 
 function htmlFromSerialAssets(
   assets: SerialAsset[],
-  { dev, template, bundleUrl }: { dev: boolean; template: string; bundleUrl?: string }
+  {
+    dev,
+    template,
+    assetPrefix,
+    bundleUrl,
+  }: {
+    dev: boolean;
+    template: string;
+    assetPrefix: string;
+    /** This is dev-only. */
+    bundleUrl?: string;
+  }
 ) {
   // Combine the CSS modules into tags that have hot refresh data attributes.
   const styleString = assets
@@ -535,8 +555,8 @@ function htmlFromSerialAssets(
         return `<style data-expo-css-hmr="${metadata.hmrId}">` + source + '\n</style>';
       } else {
         return [
-          `<link rel="preload" href="/${filename}" as="style">`,
-          `<link rel="stylesheet" href="/${filename}">`,
+          `<link rel="preload" href="${assetPrefix}/${filename}" as="style">`,
+          `<link rel="stylesheet" href="${assetPrefix}/${filename}">`,
         ].join('');
       }
     })
@@ -548,7 +568,7 @@ function htmlFromSerialAssets(
     ? `<script src="${bundleUrl}" defer></script>`
     : jsAssets
         .map(({ filename }) => {
-          return `<script src="/${filename}" defer></script>`;
+          return `<script src="${assetPrefix}/${filename}" defer></script>`;
         })
         .join('');